ENH: Modifications to pipeline to throw an exception if an invalid requested region is      detected.
ENH: Modifications to pipeline to allow it to be reset to a valid state after an
     exception is caught.

ENH: Modifications to exceptions to allow the __FILE__ and __LINE__ where the exception
     occurred to be recorded (parameters to the exception constructor).

ENH: Added Print(), PrintSelf() mechanism to exceptions.

ENH: Modified PrintSelf() for DataObject, ProcessObject, Region, ImageRegion.

ENH: Modified all instances of an ExceptionObject to be constructed with the __FILE__ and __LINE__ where the exception occurred.