(split) Create InsightApplications from Applications and Auxiliary
The InsightApplications CVS repository was created on 2003-02-19 by
copying the ,v files from the Applications and Auxiliary directories
in the main Insight repository and then "cvs rm"-ing the original
copies.  This commit was manufactured during conversion from CVS to
represent this event as a merge of the history of each directory.