ENH: Do not build SlicerExecutionModel if its path is provided
In a previous commit (5480a6b32aec3df372f08fc18882e8646ca8b8a5 [1]), External_SlicerExecutionModel.cmake has been modified and does not check if SlicerExecutionModel_DIR is defined before creating the external project to download and compile SlicerExecutionModel. Even if it is provided and if USE_SYSTEM_SLICER_EXECUTION_MODEL is set to ON, the external project for SlicerExecutionModel was still used.

[1] https://github.com/KitwareMedical/ImageViewer/commit/5480a6b32aec3df372f08fc18882e8646ca8b8a5#diff-2ead5608565906a4e7f5d574e7363861