ENH: Ruler Widget
Amended: remove using std::sqrt(), changed setIsONSDMeta to setIsONSDRuler, made ruler lines dotted.

Added a Ruler Widget that allows multiple rulers to be overlayed on an image and exported to JSON.

Added a new cClickMode CM_RULER.  Moved the screen to image index calculation in QtGlSliceView into its
own function and added the inverse and physical point transforms.

Added a new flag -R --ONSDRuler that changes the default ruler style to ONSD to the CLI app.

Added a new key input, 'u' that switches the ruler mode and ctrl-U that saves the ruler output to JSON.

Updated help information with usage.

Two modes to ruler, a "rainbow" mode that cycles between 7 colors and monotonically names each ruler and
a ONSD mode that has an "R1" and "ONSD" measure that is for the 3mm ocular orb distance and ONSD.

Rulers are placed with left-click and moved or deleted with right-click when in CM_RULER.

Use C++11 memory management features.