BUG: Overlay opacity was being overwritten - also hide select with value
OverlayOpacity was being reset when a new overlay was created. When
starting painting, it would create an overlay, and thereby reset
opacity.  So, opacity values passed on the command line were ignored.

Also, when display of "values" (details in lower right) were toggled,
the mode "select/paint/ruler/..." was still being displayed.  Now
it's display is appropriately toggled as well.